fix: updated schema for lists

+++ ui/grid/GridTileWrapper.tsx
@@ -2,6 +2,7 @@ import { useState } from 'react'
 import { YStack } from '../core/Stacks'
 import { TouchableOpacity, Pressable } from 'react-native'
 import { base } from '@/features/style'
+import { useUIStore } from '../state'
 import { router, usePathname } from 'expo-router'
 import Ionicons from '@expo/vector-icons/Ionicons'
 import { s, c } from '@/features/style'
@@ -10,16 +11,19 @@ export const GridTileWrapper = ({
   type,
   children,
   id,
+  canEdit,
   index,
   onRemove,
 }: {
   type: GridTileType
   children: React.ReactNode
+  canEdit: boolean
   id?: string
   index?: number
   onRemove?: () => void
 }) => {
   const pathname = usePathname()
+  const { editingProfile, startEditProfile, stopEditProfile } = useUIStore()
 
   let timeout: ReturnType<typeof setTimeout>
   const [actions, setActions] = useState(false)
@@ -33,26 +37,28 @@ export const GridTileWrapper = ({
     router.push(`${pathname}/details${id && `?initialId=${id}`}`)
   }
 
-  const showActions = () => {
-    clearTimeout(timeout)
-    timeout = setTimeout(() => {
-      setActions(false)
-    }, 10000)
-    setActions(true)
+  const handleLongPress = () => {
+    if (canEdit) {
+      clearTimeout(timeout)
+      timeout = setTimeout(() => {
+        setActions(false)
+      }, 10000)
+      startEditProfile()
+    }
   }
 
   return (
     <TouchableOpacity
       onPress={openDetailScreen}
-      onLongPress={showActions}
+      onLongPress={handleLongPress}
       style={[base.gridTile, specificStyles]}
     >
-      {actions && type !== 'add' && (
+      {editingProfile && type !== 'add' && (
         <YStack style={{ position: 'absolute', zIndex: 999, top: 0, right: 0 }}>
           <Pressable
             onPress={() => {
-              setActions(false)
-              onRemove()
+              stopEditProfile()
+              onRemove && onRemove()
             }}
             style={{
               transform: 'translate(8px, -8px)',
